+++ client/package.json
@@ -12,17 +12,12 @@
     "@types/react": "16.8.19",
     "@types/react-beautiful-dnd": "^11.0.2",
     "@types/react-dom": "16.8.4",
-    "@types/react-helmet": "^5.0.15",
-    "@types/react-select": "^2.0.19",
-    "darkmode-js": "^1.5.5",
     "react": "^16.8.6",
     "react-beautiful-dnd": "^11.0.4",
     "react-dnd": "^9.4.0",
     "react-dnd-html5-backend": "^9.4.0",
     "react-dom": "^16.8.6",
-    "react-helmet": "^6.0.0",
     "react-scripts": "3.0.1",
-    "react-select": "^3.0.4",
     "styled-components": "^5.1.0",
     "typescript": "3.5.1"
   },
