+++ client/src/constants/theme.ts
@@ -1,3 +1,32 @@
-export const primary: string = '#3a76f8' // CSESoc color
-export const primaryDark: string = '#084cdd' // darker variation
-export const border: string = '#bbb' // border for timetable, select, etc.
+import { createMuiTheme } from '@material-ui/core'
+
+export const darkTheme = createMuiTheme({
+    palette: {
+        type: 'dark',
+        secondary: {
+            main: "#777", // borders
+            dark: "#202020" // background
+        },
+        primary: {
+            main: "#3a76f8" // Course Select underline and NavBar
+        },
+        action: {
+            disabled: "rgba(255, 255, 255, 0.5)"
+        }
+    },
+})
+
+export const lightTheme = createMuiTheme({
+    palette: {
+        secondary: {
+            main: "#bbb", //borders
+            dark: "#fff" // background
+        },
+        primary: {
+            main: "#3a76f8" // Course Select underline and NavBar
+        },
+        action: {
+            disabled: "rgba(255, 255, 255, 0.5)"
+        }
+    }
+})
